fixed handling of pending job arrays in salljobs
pending job arrays reported
  - the CPU requested by one pending job
  - node numbers

While it is very difficult to predict the actual exact number of CPUs
that will be allocated for pending jobs due to (1) the allocation unit
[cpu, core, socket, node?] and (2) the configuration of the node the job
will be scheduled on [hyperthreading or not? cores-per-sockt?, ...],
the number for job arrays should at least be the sum of all requested
CPUs for the pending jobs in the array.

also, it doesn't make sense to report nodes for pending jobs. It is not
known ahead of time on how many nodes the jobs will end up on.

Also fixed a bug in calculating the total number of available nodes (now
considers down/drained nodes).

this fixes issue #2